utils統合
